+++ src/Request/Combination/CombinationExportRequest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace FactorioItemBrowser\Api\Client\Request\Export;
+namespace FactorioItemBrowser\Api\Client\Request\Combination;
 
 use FactorioItemBrowser\Api\Client\Request\RequestInterface;
 
 /**
- * The request of the export status.
+ * The request to export a combination.
  *
  * @author BluePsyduck <bluepsyduck@gmx.com>
  * @license http://opensource.org/licenses/GPL-3.0 GPL v3
  */
-class ExportStatusRequest implements RequestInterface
+class CombinationExportRequest implements RequestInterface
 {
 }