+++ src/Request/Combination/CombinationStatusRequest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace FactorioItemBrowser\Api\Client\Request\Export;
+namespace FactorioItemBrowser\Api\Client\Request\Combination;
 
 use FactorioItemBrowser\Api\Client\Request\RequestInterface;
 
 /**
- * The request to create a new export.
+ * The request of the combination status.
  *
  * @author BluePsyduck <bluepsyduck@gmx.com>
  * @license http://opensource.org/licenses/GPL-3.0 GPL v3
  */
-class ExportCreateRequest implements RequestInterface
+class CombinationStatusRequest implements RequestInterface
 {
 }