+++ test/serializer/Entity/ExportJobTest.php
@@ -6,7 +6,7 @@ namespace FactorioItemBrowserTestSerializer\Api\Client\Entity;
 
 use DateTime;
 use Exception;
-use FactorioItemBrowser\Api\Client\Entity\Export;
+use FactorioItemBrowser\Api\Client\Entity\ExportJob;
 use FactorioItemBrowserTestAsset\Api\Client\SerializerTestCase;
 
 /**
@@ -16,7 +16,7 @@ use FactorioItemBrowserTestAsset\Api\Client\SerializerTestCase;
  * @license http://opensource.org/licenses/GPL-3.0 GPL v3
  * @coversNothing
  */
-class ExportTest extends SerializerTestCase
+class ExportJobTest extends SerializerTestCase
 {
     /**
      * Returns the object to be serialized or deserialized.
@@ -25,11 +25,12 @@ class ExportTest extends SerializerTestCase
      */
     protected function getObject(): object
     {
-        $result = new Export();
+        $result = new ExportJob();
         $result->setStatus('abc')
                ->setCreationTime(new DateTime('2038-01-19 03:10:07'))
                ->setExportTime(new DateTime('2038-01-19 03:12:07'))
-               ->setImportTime(new DateTime('2038-01-19 03:14:07'));
+               ->setImportTime(new DateTime('2038-01-19 03:14:07'))
+               ->setErrorMessage('def');
 
         return $result;
     }
@@ -45,6 +46,7 @@ class ExportTest extends SerializerTestCase
             'creationTime' => '2038-01-19T03:10:07+00:00',
             'exportTime' => '2038-01-19T03:12:07+00:00',
             'importTime' => '2038-01-19T03:14:07+00:00',
+            'errorMessage' => 'def',
         ];
     }
 }